Modal 애니메이션 완성
